requirements.txt
openai>=1.0.0
pyyaml>=6.0
markdown>=3.4.0
